Fix handling of dynamic sized arrays
In SHDLC messages may contain dynamic arrays. For the transmit case
these dynamic array lead to runtime errors because the statically
generated length in the descriptor used for packing was not matching the
actual length.
Now the descriptor is adjusted at runtime to the actual length of the
parameter data.
When reviewing the solution that was implemented for the RX case it
turned out that always the dynamic unpack was chose due to a wrong check
in the implementation.
The new implementation will work with existing drivers and there is no
need to change the generated code.